+++ pkg/network/anonymity/anonymity_test.go
@@ -48,7 +48,7 @@ func TestError(t *testing.T) {
 func TestNodeSettings(t *testing.T) {
 	t.Parallel()
 
-	node, cancels := testNewNode(time.Minute, "", 9, 0, false)
+	node, cancels := testNewNode(time.Minute, "", 9, 0)
 	defer testFreeNodes([]INode{node}, []context.CancelFunc{cancels}, 9)
 
 	sett := node.GetSettings()
@@ -311,7 +311,7 @@ func TestEnqueuePayload(t *testing.T) {
 func TestHandleWrapper(t *testing.T) {
 	t.Parallel()
 
-	_node, cancel := testNewNode(time.Minute, "", 7, 0, true)
+	_node, cancel := testNewNode(time.Minute, "", 7, 0)
 	defer testFreeNodes([]INode{_node}, []context.CancelFunc{cancel}, 7)
 
 	node := _node.(*sNode)
@@ -319,8 +319,7 @@ func TestHandleWrapper(t *testing.T) {
 	client := node.fQueue.GetClient()
 	pubKey := client.GetPubKey()
 
-	// // ignore add public key (f2f_disabled=true)
-	// node.GetListPubKeys().AddPubKey(pubKey)
+	node.GetListPubKeys().AddPubKey(pubKey)
 
 	ctx := context.Background()
 	sett := net_message.NewSettings(&net_message.SSettings{})
@@ -438,7 +437,7 @@ func TestHandleWrapper(t *testing.T) {
 func TestStoreHashWithBroadcastMessage(t *testing.T) {
 	t.Parallel()
 
-	_node, cancel := testNewNode(time.Minute, "", 6, 0, false)
+	_node, cancel := testNewNode(time.Minute, "", 6, 0)
 	defer testFreeNodes([]INode{_node}, []context.CancelFunc{cancel}, 6)
 
 	node := _node.(*sNode)
@@ -494,7 +493,7 @@ func TestStoreHashWithBroadcastMessage(t *testing.T) {
 func TestRecvSendMessage(t *testing.T) {
 	t.Parallel()
 
-	_node, cancel := testNewNode(time.Minute, "", 5, 0, false)
+	_node, cancel := testNewNode(time.Minute, "", 5, 0)
 	defer testFreeNodes([]INode{_node}, []context.CancelFunc{cancel}, 5)
 
 	ctx, cancel := context.WithCancel(context.Background())
@@ -572,7 +571,7 @@ func testNewNodes(t *testing.T, timeWait time.Duration, addresses [2]string, typ
 	addrs := [5]string{"", "", addresses[0], "", addresses[1]}
 
 	for i := 0; i < 5; i++ {
-		nodes[i], cancels[i] = testNewNode(timeWait, addrs[i], typeDB, i, false)
+		nodes[i], cancels[i] = testNewNode(timeWait, addrs[i], typeDB, i)
 		if nodes[i] == nil {
 			t.Errorf("node (%d) is not running %d", i, typeDB)
 			return [5]INode{}, [5]context.CancelFunc{}
@@ -660,7 +659,7 @@ func (p *stLogging) HasErro() bool {
 }
 */
 
-func testNewNode(timeWait time.Duration, addr string, typeDB, numDB int, f2fDisabled bool) (INode, context.CancelFunc) {
+func testNewNode(timeWait time.Duration, addr string, typeDB, numDB int) (INode, context.CancelFunc) {
 	db, err := database.NewKVDatabase(
 		database.NewSettings(&database.SSettings{
 			FPath:     fmt.Sprintf(tcPathDBTemplate, typeDB, numDB),
@@ -678,7 +677,6 @@ func testNewNode(timeWait time.Duration, addr string, typeDB, numDB int, f2fDisa
 	node := NewNode(
 		NewSettings(&SSettings{
 			FServiceName:  "TEST",
-			FF2FDisabled:  f2fDisabled,
 			FNetworkMask:  networkMask,
 			FFetchTimeout: timeWait,
 		}),