+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - Update `cmd/tools/pmanager`: delete constant from crypto/symmetric package, update keySize 32 -> 33 bytes
 - Update `pkg/network/conn`: add context to NewConn function, rename NewConn -> Connect
+- Update `pkg/network/anonymity`: delete f2f_disabled=true option (insecure)
+- Update `pkg/network/anonymity/queue`: delete queue_period=0 option (insecure)
 
 <!-- ... -->